c# fram timeout 单位改成milliseconds

+++ llbc/src/comm/Service.cpp
@@ -1586,7 +1586,7 @@ void LLBC_Service::HandleQueuedEvents()
         if (_frameTimeout == LLBC_INFINITE)
             continue;
 
-        if(UNLIKELY((LLBC_CPUTime::Current() - _begSvcTime).ToNanoSeconds() >= _frameTimeout))
+        if(UNLIKELY(static_cast<uint64>((LLBC_CPUTime::Current() - _begSvcTime).ToNanoSeconds()) >= _frameTimeout))
             break;
         #endif  // LLBC_CFG_COMM_ENABLE_SERVICE_FRAME_TIMEOUT
     }
